% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfafunctions.R
\name{collate_samples}
\alias{collate_samples}
\title{collate_samples: takes a matrix of consumption values and re-formats
it to give a dataframe in Tidy format containing Line, Group, Rep, and
consumption values. Sample replicates are numbered down columns;
e.g. A1, B1, C1, D1 -> 1, 2, 3, 4}
\usage{
collate_samples(
  con_matrix,
  matrix_type = "cols12",
  layout_matrix = NULL,
  ID = "PROJ_LINE_S_ASSY"
)
}
\arguments{
\item{con_matrix}{A matrix containing consumption values
in a wellplate format.

vol The initial volume dispensed into the wells of the microplate.
Default vol = 10.

matrix_type A character variable indicating the type of built-in
layout used to organize the 96-well culture plate. Default = "cols12".
See collate_samples() function for details about layouts.

For custom matrix layouts, matrix_type = "custom".
layout_matrix A matrix containing a user-provided layour matrix for
sample in the 96-well culture plate. Default layout_matrix = NULL.
matrix_type must be matrix_type = "custom".

ID A character string containing the sample ID for the plate.}
}
\description{
collate_samples: takes a matrix of consumption values and re-formats
it to give a dataframe in Tidy format containing Line, Group, Rep, and
consumption values. Sample replicates are numbered down columns;
e.g. A1, B1, C1, D1 -> 1, 2, 3, 4
}
\examples{
imp_xml = xml2::read_xml(xml_doc_path)

abs_reads = read_plates(imp_xml)

con_data = consumption_from_array(abs_reads)

Output:
Line   Group   Rep    Cons
0000   M       1      0.11549404
0000   M       2      0.00803428
0000   M       3      0.08090379
0000   F       1      0.19400821
0000   G       1      0.06189788
0000   G       2     -0.07962330
}
\keyword{}
