% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfafunctions.R
\name{con_time}
\alias{con_time}
\title{con_time: estimates the consumption time and matches the plate IDs
from internal xml document.}
\usage{
con_time(imp_xml)
}
\arguments{
\item{imp_xml}{An internal xml document imported via xml2::read_xml().}
}
\description{
con_time: estimates the consumption time and matches the plate IDs
from internal xml document.
}
\examples{
imp_xml = xml2::read_xml(xml_doc_path)

# Get all reads for double checking
all_reads = import_plates(imp_xml)

print(all_reads)

Output:
ID                  Project  Line   Sex  Assay   ReadTime
HARD_0000_0_SPEC    HARD     0000   0    SPEC    2020-07-02 11:54:00
HARD_0000_0_SPEC    HARD     0000   0    SPEC    2020-07-03 08:49:00
HARD_TAPE_0_SPEC    HARD     TAPE   0    SPEC    2020-07-02 14:10:00
HARD_TAPE_0_SPEC    HARD     TAPE   0    SPEC    2020-07-03 10:53:00

# Calculate consumption times from matching IDs
consumption_times = con_time(imp_xml)

print(consumption_times)

ID                  Project  Line   Sex  Assay   Duration(hr)
HARD_0000_0_SPEC    HARD     0000   0    SPEC    20.9166666666667
HARD_TAPE_0_SPEC    HARD     TAPE   0    SPEC    20.7166666666667

}
\keyword{consumption}
\keyword{elaspsed}
\keyword{time}
\keyword{time,}
