% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfafunctions.R
\name{consumption_formula}
\alias{consumption_formula}
\title{consumption_formula: formula used to calculate consumption}
\usage{
consumption_formula(vol, plate1, plate2)
}
\arguments{
\item{vol}{The initial volume dispensed into the wells of the microplate.
Default vol = 10.

plate1 The pre-exposure plate in the calculation.

plate2 The post-exposure plate in the calculation.}
}
\description{
consumption_formula: formula used to calculate consumption
}
\keyword{calculate}
\keyword{formula,}
