% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfafunctions.R
\name{find_wells}
\alias{find_wells}
\title{find_wells: blank the plates and identify wells of interest based
on the specified absorbance cutoff}
\usage{
find_wells(readings, cutoff = 1, bin = TRUE, blank = TRUE)
}
\arguments{
\item{readings}{A matrix containing the absorbance readings.
cutoff The absorbance threshold for identifying wells of interest.
Full wells tend to have absorbances of ~2.2. Default cutoff = 1.0.
bin A logical variable indicating whether to zero the wells below
the threshold value. Default bin = TRUE.
blank A logical variable indicating whether to substract the
background absorbance from each reading. Default = TRUE.}
}
\description{
find_wells: blank the plates and identify wells of interest based
on the specified absorbance cutoff
}
\keyword{absorbance,}
\keyword{blank}
\keyword{cutoff,}
\keyword{read,}
\keyword{threshold,}
