% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfafunctions.R
\name{import_plates}
\alias{import_plates}
\title{import_plates: generates a full readout of all plate reads in an
xml document. Calls parse_IDs and parse_time functions to reformat
the xml attributes.}
\usage{
import_plates(imp_xml)
}
\arguments{
\item{imp_xml}{An internal xml document imported via xml2::read_xml().}
}
\description{
import_plates: generates a full readout of all plate reads in an
xml document. Calls parse_IDs and parse_time functions to reformat
the xml attributes.
}
\examples{
imp_xml = xml2::read_xml(xml_doc_path)
all_reads = import_plates(imp_xml)
print(all_reads)

Output:
ID                  Project Line   Sex  Assay    ReadTime
HARD_0420_M_BLZR    HARD    0420   M    BLZR     2020-07-02 11:54:00
TIPS_8008_F_SPEC    TIPS    8008   F    SPEC     2020-07-03 08:49:00
TORK_TAPE_F_SPEC    TORK    TAPE   F    SPEC     2020-07-02 14:10:00
HARD_BAPE_M_SPEC    HARD    BAPE   M    SPEC     2020-07-03 10:53:00
}
\keyword{import,}
\keyword{information}
\keyword{plates,}
\keyword{read}
