% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfafunctions.R
\name{make_1536}
\alias{make_1536}
\title{make_1536: re-formats the plates from 384 with 4 sub-areas to 1536}
\usage{
make_1536(imp_xml, plate_node)
}
\arguments{
\item{imp_xml}{An internal xml document imported via xml2::read_xml().
plate_node The location of the node containing the plate reads that
are being re-formatted. Note: For clarification concerning navigating
xml documents, search for XPATH tutorials.}
}
\description{
make_1536: re-formats the plates from 384 with 4 sub-areas to 1536
}
\examples{
imp_xml = xml2::read_xml(xml_doc_path)
plate_node = "/Experiment/PlateSections[2]/PlateSection"

# Get a 32x48 matrix of absorbance readings
abs_1536 = make_1536(imp_xml, plate_node)

print(abs_1536)

Output:
     [,1]   [,2]   [,3]   [,4]   ...[,48]
[1,] 0.2774 0.1596 0.1390 0.4679
[2,] 0.1667 0.1481 0.3128 0.1542
[3,] 0.1526 0.2950 2.2933 0.1468
[4,] 0.1509 0.1556 0.1655 0.2997
...
[32,]
}
\keyword{1536,}
\keyword{384}
\keyword{format,}
