% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfafunctions.R
\name{parse_IDs}
\alias{parse_IDs}
\title{parse_IDs: converts a vector of sample IDS into a matrix of
factors for each respective ID.}
\usage{
parse_IDs(id, sep = "_")
}
\arguments{
\item{id}{A vector containing character strings of sample IDs
with each factor separated by a character. sep The character
separating factors within the sample ID. Default sep = "_".}
}
\description{
parse_IDs: converts a vector of sample IDS into a matrix of
factors for each respective ID.
}
\examples{
sampleID = c("DGRP_0120_M_MUSH",
             "DGRP_5430_F_BLIP")
factors = parse_IDs(sampleID)
print(factors)

Output:
     [,1]                [,2]    [,3]    [,4] [,5]
[1,] "DGRP_0120_M_MUSH", "DGRP", "0120", "M", "MUSH"
}
\keyword{IDs,}
\keyword{factors}
\keyword{samples,}
