% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfafunctions.R
\name{parse_time}
\alias{parse_time}
\title{parse_time: converts date-time to lubridate format for easier
time calculations}
\usage{
parse_time(date_time_string)
}
\arguments{
\item{date_time_string}{A character string containing the
date and time as exported from Molecular Devices iD5 plate reader.
i.e. HH:MM am/pm M/D/YYYY}
}
\description{
parse_time: converts date-time to lubridate format for easier
time calculations
}
\examples{
ReadTime = "2:54 PM 7/2/2020"
time = parse_time(ReadTime)
print(time)

Output: 2020-07-02 14:54:00
}
\keyword{date}
\keyword{time,}
