% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfafunctions.R
\name{read_plates}
\alias{read_plates}
\title{read_plates: creates a 3 dimensional array containing the plate reads
for each sample in a \link{plate row, plate column, plate ID} format}
\usage{
read_plates(imp_xml, cutoff = 1, blank = TRUE)
}
\arguments{
\item{imp_xml}{An internal xml document imported via xml2::read_xml().
cutoff The absorbance threshold for identifying wells of interest.
Full wells tend to have absorbances of ~2.2. Default cutoff = 1.0.
blank A logical variable indicating whether to substract the
background absorbance from each reading. Default = TRUE.}
}
\description{
read_plates: creates a 3 dimensional array containing the plate reads
for each sample in a \link{plate row, plate column, plate ID} format
}
\examples{
imp_xml = xml2::read_xml(xml_doc_path)

abs_reads = read_plates(imp_xml)

print(abs_reads)

Output:
     [,1]   [,2]   [,3]   [,4]   ...[,12]
[1,] 0.2774 0.1596 0.1390 0.4679
[2,] 0.1667 0.1481 0.3128 0.1542
[3,] 0.1526 0.2950 2.2933 0.1468
[4,] 0.1509 0.1556 0.1655 0.2997
...
[8,]
}
\keyword{absorbance,}
\keyword{blank}
\keyword{cutoff,}
\keyword{read,}
\keyword{threshold,}
