% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfafunctions.R
\name{get_stats}
\alias{get_stats}
\title{get_stats: apply a standard set of statistical tests to the data}
\usage{
get_stats(con_list, keep_row_names = TRUE)
}
\arguments{
\item{con_list}{A list of consumption values belonging to a single
group for analysis.}
}
\description{
get_stats: apply a standard set of statistical tests to the data
}
\keyword{descriptive}
\keyword{normality,}
\keyword{statistics,}
